// src/ussd/dto/ussd.dto.ts
export class UssdDto {
  userId?: string;
  msisdn?: string;
  phoneNumber?: string;
  userData: string;
  msgType: any;
  network?: string;
  sessionId?: any;
  transId?: any;
  msg?: string;
  USERID?: string;
  MSISDN: string;
  USERDATA?: string;
  MSGTYPE?: any;
  NETWORK?: string;
  MSG: string;
  SESSIONID: string;

}